dashboard_project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apitools==0.1.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
Django==4.0
django-allauth==0.59.0
djangorestframework==3.14.0
drf-yasg==1.21.7
fasteners==0.19
frozenlist==1.4.0
google-apitools==0.5.32
google-auth==2.23.2
google-auth-oauthlib==1.1.0
google-businesscommunications==2.1.0
google-businessmessages==1.0.5
httplib2==0.22.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
mysqlclient==2.2.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
whitenoise==6.6.0
yarl==1.9.2
